+++ src/main/java/foodscooter/api/controllers/RiderController.java
@@ -1,13 +1,12 @@
 package foodscooter.api.controllers;
 
-import foodscooter.model.Order;
-import foodscooter.model.rider.FullTimeSchedule;
-import foodscooter.model.rider.PartTimeShift;
-import foodscooter.model.rider.RiderType;
-import foodscooter.model.rider.SalaryInfo;
-import foodscooter.model.users.Rider;
+import foodscooter.model.orders.Order;
+import foodscooter.model.users.rider.FullTimeSchedule;
+import foodscooter.model.users.rider.PartTimeShift;
+import foodscooter.model.users.rider.RiderType;
+import foodscooter.model.users.rider.SalaryInfo;
+import foodscooter.model.users.rider.Rider;
 import foodscooter.repositories.JdbcRidersRepository;
-import foodscooter.repositories.JdbcUsersRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
