+++ src/main/java/foodscooter/api/dtos/login/Credentials.java
@@ -4,11 +4,6 @@ public class Credentials {
   private String username;
   private String password;
 
-  public Credentials(String username, String password) {
-    this.username = username;
-    this.password = password;
-  }
-
   public String getUsername() {
     return username;
   }
