--- src/main/java/foodscooter/model/Promotion.java
@@ -1,4 +1,0 @@
-package foodscooter.model;
-
-public class Promotion {
-}
