+++ src/main/java/foodscooter/model/restaurants/Promotion.java
@@ -1,0 +1,4 @@
+package foodscooter.model.restaurants;
+
+public class Promotion {
+}
