+++ src/main/java/foodscooter/model/restaurants/Restaurant.java
@@ -1,4 +1,4 @@
-package foodscooter.model;
+package foodscooter.model.restaurants;
 
 public class Restaurant {
   private int id;
