+++ src/main/java/foodscooter/model/users/UserType.java
@@ -1,4 +1,4 @@
-package foodscooter.model;
+package foodscooter.model.users;
 
 public enum UserType {
   DELIVERY_RIDER("Delivery Rider"),
