+++ src/main/java/foodscooter/model/users/rider/FullTimeSchedule.java
@@ -1,4 +1,4 @@
-package foodscooter.model.rider;
+package foodscooter.model.users.rider;
 
 public class FullTimeSchedule {
   private int dayOption;
