+++ src/main/java/foodscooter/model/users/rider/PartTimeShift.java
@@ -1,4 +1,4 @@
-package foodscooter.model.rider;
+package foodscooter.model.users.rider;
 
 public class PartTimeShift {
   private int ptsid;
