+++ src/main/java/foodscooter/model/users/rider/RiderType.java
@@ -1,4 +1,4 @@
-package foodscooter.model.rider;
+package foodscooter.model.users.rider;
 
 public enum RiderType {
   FULL_TIME,
