+++ src/main/java/foodscooter/model/users/rider/SalaryInfo.java
@@ -1,4 +1,4 @@
-package foodscooter.model.rider;
+package foodscooter.model.users.rider;
 
 public class SalaryInfo {
   private int numOfOrder;
