+++ src/main/java/foodscooter/repositories/JdbcRidersRepository.java
@@ -1,11 +1,11 @@
 package foodscooter.repositories;
 
-import foodscooter.model.Order;
-import foodscooter.model.rider.FullTimeSchedule;
-import foodscooter.model.rider.PartTimeShift;
-import foodscooter.model.rider.RiderType;
-import foodscooter.model.rider.SalaryInfo;
-import foodscooter.model.users.Rider;
+import foodscooter.model.orders.Order;
+import foodscooter.model.users.rider.FullTimeSchedule;
+import foodscooter.model.users.rider.PartTimeShift;
+import foodscooter.model.users.rider.RiderType;
+import foodscooter.model.users.rider.SalaryInfo;
+import foodscooter.model.users.rider.Rider;
 import foodscooter.repositories.specifications.RidersRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
